requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
sqlparse==0.4.2
asgiref==3.5.2
backports.zoneinfo==0.2.1
cffi==1.15.1
cryptography==38.0.1
Django==4.0.6
django-rest-knox==4.2.0
djangorestframework==3.14.0
Pillow==9.2.0
psycopg2==2.9.3
pycparser==2.21
pytz==2022.4
sqlparse==0.4.2
